add rangeslidervalue
